+++ example/lib/theme.dart
@@ -1,0 +1,10 @@
+import 'package:fluent_ui/fluent_ui.dart';
+
+class AppTheme extends ChangeNotifier {
+  ThemeMode _mode = ThemeMode.system;
+  ThemeMode get mode => _mode;
+  set mode(ThemeMode mode) {
+    _mode = mode;
+    notifyListeners();
+  }
+}